all sadmin and cw issues done

+++ app/Http/Controllers/API/CompanyController.php
@@ -10,7 +10,7 @@ use App\Models\PorjectCompanyWorker;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
-
+use Validator;
 class CompanyController extends Controller
 {
     /**
@@ -26,7 +26,7 @@ class CompanyController extends Controller
             foreach ($companies as $key => $value) {
                 $value->image=$img_path.'/'.$value->image;
             }
-           //dd($companies);
+            //dd($companies);
             return response()->json([
                 $companies
             ], 200);
@@ -41,17 +41,17 @@ class CompanyController extends Controller
         try {
             //dd($pid);
             $companies = Company::all();
-           $project_company = PorjectCompanyWorker::where('project_id', $pid)->get();
-           $company_ids= array();
-           foreach($project_company as $pc){
-               $company_ids[] = $pc->company_worker_id;
-           }
-           // dd($company_ids);
-           $companies = Company::whereNotIn('id', $company_ids)->get();
+            $project_company = PorjectCompanyWorker::where('project_id', $pid)->get();
+            $company_ids= array();
+            foreach($project_company as $pc){
+                $company_ids[] = $pc->company_worker_id;
+            }
+            // dd($company_ids);
+            $companies = Company::whereNotIn('id', $company_ids)->get();
             $img_path=asset('company_images/');
-           foreach ($companies as $cmp){
-               $cmp->image=$img_path.'/'.$cmp->image;
-           }
+            foreach ($companies as $cmp){
+                $cmp->image=$img_path.'/'.$cmp->image;
+            }
             return response()->json([
                 $companies
             ], 200);
@@ -92,26 +92,34 @@ class CompanyController extends Controller
             $company->manager_name = $request->manager_name ;
             $company->manager_email = $request->manager_email ;
             $company->manager_phone = $request->manager_phone ;
-            //return $this->responseSuccess($step);
+
+           /* $input['email'] = $request->email;
+            $rules = array('email' => 'unique:users,email');
+            $validator = Validator::make($input, $rules);
+            if ($validator->fails()) {
+                dd('email exists already');
+            }*/
+            $user = new User();
+
+            $user->email = $request->email;
+            $user->first_name = $request->name;
+            $user->password = Hash::make($request->password);
+
             if (!empty($request->image)) {
                 $imageName = time() . '.' . $request->image->extension();
                 $request->image->move(public_path('company_images'), $imageName);
                 $company->image = $imageName;
+                $user->img = 'dummy_image.png';
             }else{
                 $company->image = 'dummy_image.png';
+                $user->img = 'dummy_image.png';
+
             }
 
-            //dd($company);
             $company->save();
-           // dd($request->all());
-
-            $user = new User();
-            $user->email = $request->email;
-            $user->first_name = $request->name;
-            $user->password = Hash::make($request->password);
             $user->company = $company->id;
             $user->save();
-           // dd($company->id);
+            // dd($company->id);
             return response()->json([
                 $company
             ], 200);
@@ -161,59 +169,61 @@ class CompanyController extends Controller
     public function update(Request $request, Company $company)
     {
         try{
-
             if(!empty($request->name))
-                //dd($request->name);
-            $company->name = $request->name ;
+                $company->name = $request->name ;
 
             if(!empty($request->organization_number))
-            $company->organization_number = $request->organization_number ;
+                $company->organization_number = $request->organization_number ;
 
             if(!empty($request->address))
-            $company->address = $request->address ;
+                $company->address = $request->address ;
 
             if(!empty($request->contact_number))
-            $company->contact_number = $request->contact_number ;
+                $company->contact_number = $request->contact_number ;
 
             if(!empty($request->email))
-            $company->email = $request->email ;
+                $company->email = $request->email ;
 
             if(!empty($request->password))
-            $company->password= Hash::make($request->password);
+                $company->password= Hash::make($request->password);
 
             if(!empty($request->confirm_password))
-            $company->confirm_password = Hash::make($request->confirm_password);
+                $company->confirm_password = Hash::make($request->confirm_password);
 
             if(!empty($request->image))
-            $company->image = $request->image ;
+                $company->image = $request->image ;
 
             if(!empty($request->manager_name))
-            $company->manager_name = $request->manager_name ;
+                $company->manager_name = $request->manager_name ;
 
             if(!empty($request->manager_email))
-            $company->manager_email = $request->manager_email ;
+                $company->manager_email = $request->manager_email ;
 
             if(!empty($request->manager_phone))
-            $company->manager_phone = $request->manager_phone ;
+                $company->manager_phone = $request->manager_phone ;
 
-           // dd($request->name);
-            //return $this->responseSuccess($step);
-            if (!empty($request->image)) {
-                $imageName = time() . '.' . $request->image->extension();
-                $request->image->move(public_path('company_images'), $imageName);
+
+            $user = User::where('company', $company->id)->first();
+
+            if(!empty($request->password))
+                $user->password = Hash::make($request->password);
+            if(!empty($request->name))
+                $user->first_name = $request->name;
+            if(!empty($request->email))
+                $user->email = $request->email ;
+
+            if (!empty($request->image)){
+                $base64_image = $request->image; // your base64 encoded
+                @list($type, $file_data) = explode(';', $base64_image);
+                @list(, $file_data) = explode(',', $file_data);
+                $imageName = time().'.'.'png';
+                \File::put(public_path('company_images/').$imageName, base64_decode($file_data));
                 $company->image = $imageName;
             }
 
-            //dd($company->id);
-            $company->update();
-            $user = User::where('company', $company->id)->first();
-            if(!empty($request->password))
-            $user->password = Hash::make($request->password);
-            if(!empty($request->name))
-            $user->first_name = $request->name;
+            $company->save();
+            $user->save();
 
-            $user->update();
-          //  dd($user);
 
             return response()->json([
                 $company
@@ -234,8 +244,10 @@ class CompanyController extends Controller
     {
         try{
             $company->delete();
+
             $user = User::where('company', $company->id)->first();
-            $user->delete();
+            if($user != NULL)
+                $user->delete();
 
             return response()->json([
                 'deleted'
